[Fix] Fix while_op to work correctly
Fix while_op and modify unittest for the op.
We expect an executable array s2 in another executable array like { 1 { s2 } }
is pushed to the stack as an ELEMENT_EXECUTABLE_ARRAY.
This is because s2 might be used with a control op:
e.g., 1 { s2 } if.
Previous version that evaluates the inside of a nested executable array at once
was wrong; this commit fixes it.